Updated code to new server. Updated some services and model. Added model classes.

+++ Source/SnapSafari/app/src/main/java/ua/in/zeusapps/snapsafari/models/EventRequest.java
@@ -20,7 +20,7 @@ public class EventRequest {
         _radius = radius;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return _latitude;
     }
 
@@ -28,7 +28,7 @@ public class EventRequest {
         _latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return _longitude;
     }
